requirements.txt
django == 5.1.2
djangorestframework == 3.15.2
drf-yasg == 1.21.7
pytz == 2024.2
setuptools == 75.1.0
django-multiselectfield == 0.1.13